requirements.txt
--extra-index-url https://erc.timetoact-group.at/
erc3>=1.2.0
openai>=2.8.1
pydantic>=2.0.0
python-dotenv>=1.0.0

